[J:MEP-0] Move JVM ImageLoader initializer into lib/design/fe
Coil's `AsyncImage` is part of the design system, and `lib/design/fe`
already owns its sibling `JvmDesignInitializer` (FileKit setup). Moving
the image-loader wiring next to it keeps platform-specific design
initialization in one place and removes the need for a per-platform
`platformModule` split inside `composeApp`.

- Add `JvmImageLoaderInitializer : ComposeInitializer` in
  `lib/design/fe/src/jvmMain/.../initializers/`, bound via the existing
  `DesignModule.jvm.kt` actual.
- Add the JVM-only `lib.storage.fe.api` and `ktor-client-okhttp` deps to
  `lib/design/fe`, where they actually live now.
- Revert the composeApp per-platform `platformModule` actuals — the
  shared `nonWebMain` empty module is back, since composeApp itself no
  longer contributes JVM-specific bindings.
