--- composeApp/src/jvmMain/kotlin/cz/adamec/timotej/snag/di/AppModule.jvm.kt
@@ -1,25 +1,0 @@
-/*
- * Copyright (c) 2026 Timotej Adamec
- * SPDX-License-Identifier: MIT
- *
- * This file is part of the thesis:
- * "Multiplatform snagging system with code sharing maximisation"
- *
- * Czech Technical University in Prague
- * Faculty of Information Technology
- * Department of Software Engineering
- */
-
-package cz.adamec.timotej.snag.di
-
-import cz.adamec.timotej.snag.initializer.SnagDesktopImageLoaderInitializer
-import cz.adamec.timotej.snag.lib.design.fe.initializer.ComposeInitializer
-import org.koin.core.module.Module
-import org.koin.core.module.dsl.singleOf
-import org.koin.dsl.bind
-import org.koin.dsl.module
-
-internal actual val platformModule: Module =
-    module {
-        singleOf(::SnagDesktopImageLoaderInitializer) bind ComposeInitializer::class
-    }
